+++ src/status/AttackStatus.java
@@ -3,7 +3,7 @@ package status;
 public class AttackStatus extends Status {
     public String name;
 
-    public AttackStatus(int value) {
+    public AttackStatus(double value) {
         super(value);
         this.name = "attackStatus";
     }
