+++ src/status/DefenseStatus.java
@@ -3,7 +3,7 @@ package status;
 public class DefenseStatus extends Status {
     public String name;
 
-    public DefenseStatus(int value) {
+    public DefenseStatus(double value) {
         super(value);
         this.name = "defenseStatus";
     }
