+++ src/status/SpeedStatus.java
@@ -3,7 +3,7 @@ package status;
 public class SpeedStatus extends Status {
     public String name;
 
-    public SpeedStatus(int value) {
+    public SpeedStatus(double value) {
         super(value);
         this.name = "speedStatus";
     }
